Add more Apache Kafka examples and utilities.
