+++ temporal-data-and-realtime-algorithm/temporal-apache-kafka/documents/config/server.properties
@@ -43,9 +43,11 @@ advertised.host.name=localhost
 advertised.port=9092
 
 # The number of threads handling network requests
+# 处理网络请求的线程数量
 num.network.threads=3
 
 # The number of threads doing disk I/O
+# 处理磁盘IO的线程数量
 num.io.threads=8
 
 # The send buffer (SO_SNDBUF) used by the socket server
@@ -55,6 +57,7 @@ socket.send.buffer.bytes=102400
 socket.receive.buffer.bytes=102400
 
 # The maximum size of a request that the socket server will accept (protection against OOM)
+# 请求的最大大小
 socket.request.max.bytes=104857600
 
 
@@ -63,7 +66,7 @@ socket.request.max.bytes=104857600
 # A comma seperated list of directories under which to store log files
 # 日志文件
 #log.dirs=/tmp/kafka-logs
-log.dirs=/Users/jiedong/software/kafka_2.11-0.9.0.1/data/kafka-logs
+log.dirs=/Users/zhang/Documents/software/kafka_2.11-0.9.0.1/data/kafka-logs
 
 # The default number of log partitions per topic. More partitions allow greater
 # parallelism for consumption, but this will also result in more files across
@@ -74,6 +77,7 @@ num.partitions=1
 
 # The number of threads per data directory to be used for log recovery at startup and flushing at shutdown.
 # This value is recommended to be increased for installations with data dirs located in RAID array.
+# 每个数据目录用于处理启动时日志恢复和关闭时刷出的线程数量
 num.recovery.threads.per.data.dir=1
 
 ############################# Log Flush Policy #############################
