+++ temporal-data-and-realtime-algorithm/temporal-apache-kafka/documents/config/zookeeper.properties
@@ -15,7 +15,7 @@
 # the directory where the snapshot is stored.
 # 快照存放目录
 #dataDir=/tmp/zookeeper
-dataDir=/Users/jiedong/software/kafka_2.11-0.9.0.1/data/zookeeper
+dataDir=/Users/zhang/Documents/software/kafka_2.11-0.9.0.1/data/zookeeper
 # the port at which the clients will connect
 # 暴露给客户端连接的端口
 #clientPort=2181
