+++ temporal-data-and-realtime-algorithm/temporal-apache-kafka/src/main/java/com/spike/giantdataanalysis/kafka/support/package-info.java
@@ -1,0 +1,13 @@
+/**
+ * <pre>
+ * Kafka配置常量
+ * 
+ * Broker: http://kafka.apache.org/documentation.html#brokerconfigs
+ * Consumer: http://kafka.apache.org/documentation.html#newconsumerconfigs
+ * Connect: http://kafka.apache.org/documentation.html#connectconfigs
+ * Producer: http://kafka.apache.org/documentation.html#producerconfigs
+ * Topic: http://kafka.apache.org/documentation.html#topic-config
+ * </pre>
+ * @author zhoujiagen
+ */
+package com.spike.giantdataanalysis.kafka.support;
