+++ temporal-data-and-realtime-algorithm/temporal-apache-kafka/src/main/resources/log4j.properties
@@ -1,18 +1,17 @@
 #log4j.rootLogger=debug, console, file, socket
 log4j.rootLogger=debug, console, file
 
-log4j.logger.com.spike.giantdataanalysis=DEBUG, kafka
-
-### kafka
-log4j.appender.kafka=org.apache.kafka.log4jappender.KafkaLog4jAppender
-log4j.appender.kafka.topic=log-analysis
-log4j.appender.kafka.brokerList=localhost:9092
-log4j.appender.kafka.compressionType=none
-log4j.appender.kafka.requiredNumAcks=0
-log4j.appender.kafka.syncSend=true
-log4j.appender.kafka.layout=org.apache.log4j.PatternLayout
-log4j.appender.kafka.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss,SSS} [%p] [%t] %l => %m%n
-
+#log4j.logger.com.spike.giantdataanalysis=DEBUG, kafka
+#
+#### kafka
+#log4j.appender.kafka=org.apache.kafka.log4jappender.KafkaLog4jAppender
+#log4j.appender.kafka.topic=log-analysis
+#log4j.appender.kafka.brokerList=localhost:9092
+#log4j.appender.kafka.compressionType=none
+#log4j.appender.kafka.requiredNumAcks=0
+#log4j.appender.kafka.syncSend=true
+#log4j.appender.kafka.layout=org.apache.log4j.PatternLayout
+#log4j.appender.kafka.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss,SSS} [%p] [%t] %l => %m%n
 
 ### ConsoleAppender
 log4j.appender.console=org.apache.log4j.ConsoleAppender
@@ -21,9 +20,6 @@ log4j.appender.console.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss,SSS} [%p]
 
 ### RollingFileAppender
 log4j.appender.file=org.apache.log4j.RollingFileAppender
-# log at work
-#log4j.appender.file.File=/Users/jiedong/logs/gda-apache-kafka.log
-# log at home
 log4j.appender.file.File=/Users/zhang/logs/gda-apache-storm.log
 #log4j.appender.file.MaxFileSize=10MB
 #log4j.appender.file.MaxBackupIndex=1
